+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Mulberry_Warranty changelog
 ========================
 
+1.4.0:
+- Added logic to init inline/modal offers when there were no offers for the initial product selection
+- Update frontend logic to use new SDK modal/inline offer checks
+
 1.3.4:
 - Added missing "self" declaration for the PDP component